Added -g grouping option which groups models together with there respective application inside the visualisation. Fixed problem with -a (autodetect all applications) and empty nodes.
The latter required some pretty big changes, but from my (still limited) understanding of the dot format this is a bit cleaner then before as it generated all model nodes first and relations after that. Because of this generation of the dot file now happens in several stages. Which was also needed to be able to skip generating relation nodes when they already exists as a full model node as you need to have walked down all the models to know if the model name a relation points to already exists or not.

git-svn-id: https://django-command-extensions.googlecode.com/svn/trunk@19 70537768-783f-0410-b1a0-8399f7a62752